@@ -1,7 +1,7 @@
 <Project Sdk="Microsoft.NET.Sdk">
 
     <PropertyGroup>
-        <TargetFramework>net5.0</TargetFramework>
+        <TargetFramework>net6.0</TargetFramework>
         <Title>JuliadotNET</Title>
         <Authors>Johnathan Bizzano</Authors>
         <PackageProjectUrl>https://github.com/HyperSphereStudio/JULIAdotNET</PackageProjectUrl>
@@ -26,7 +26,7 @@
     </ItemGroup>
 
     <ItemGroup>
-      <PackageReference Include="Lokad.ILPack" Version="0.2.0" />
+      <PackageReference Include="Lokad.ILPack" Version="0.1.7" />
     </ItemGroup>
 
 </Project>
